updating files structure
