+++ scripts/editor/selectors.js
@@ -1,6 +1,59 @@
 import classnames from 'classnames';
 
 /**
+ * Returns BEM selector for html class and check if Condition part is set.
+ *
+ * @param {boolean} condition - Check condition.
+ * @param {string} block      - BEM Block selector.
+ * @param {string} [element]  - BEM Element selector.
+ * @param {string} [modifier] - BEM Modifier selector.
+ *
+ * @access public
+ *
+ * @return string
+ *
+ * Usage:
+ * ```js
+ * selector(!(condition1 && condition2), 'foo-boo');
+ *
+ * selector(condition, 'foo', 'boo');
+ *
+ * selector(!condition, 'foo', 'boo', 'poo');
+ * ```
+ *
+ * Equivalent:
+ * ```js
+ * !(condition1 && condition2) ? 'foo-boo' : '';
+ *
+ * condition ? 'foo__boo' : '';
+ *
+ * !(condition) ? 'foo__boo--poo' : '';
+ * ```
+ *
+ * Output:
+ * ```js
+ * foo-boo
+ *
+ * foo__boo
+ *
+ * foo__boo--poo
+ */
+export const selector = (condition, block, element = '', modifier = '') => {
+	let fullModifier = '';
+	let fullElement = '';
+
+	if (element) {
+		fullElement = `__${element}`;
+	}
+
+	if (modifier) {
+		fullModifier = `--${modifier}`;
+	}
+
+	return condition ? `${block}${fullElement}${fullModifier}` : '';
+};
+
+/**
  * Create responsive selectors used for responsive attributes.
  *
  * @param {array}   items              - Array of breakpoints.
@@ -41,3 +94,4 @@ export const responsiveSelectors = (items, selector, parent, useModifier = true)
 
 	return classnames(output);
 };
+
