+++ scripts/helpers/index.js
@@ -1,11 +1,6 @@
 // All exports are sorted in alphabetical order.
 
 export { camelize } from './camelize';
-export {
-	checkAttr,
-	checkAttrResponsive,
-	getAttrKey
-} from './check-attr';
 export { cookies } from './cookies';
 export { debounce } from './debounce';
 export { throttle } from './throttle';
@@ -14,9 +9,7 @@ export { dynamicImport } from './dynamic-import';
 export { elementChildrenHeight } from './element-children-height';
 export { escapeString } from './escape-string';
 export { getNavigatorVibrate } from './navigator';
-export { responsiveSelectors } from './responsive-selectors';
-export { selector } from './selector';
 export {
 	truncateMiddle,
-	unescapeHTML
+	unescapeHTML,
 } from './text-helpers';
